update styling cost controller add cost modal
