+++ trip_calculator/trip_calculator/templates/trip_calculator/home_table/account_edit_modal.html
@@ -21,28 +21,28 @@
                             <div class="container-fluid pt-3">
                                 <div class="row justify-content-center align-items-center">
                                     <div class="col-12 col-md-6 col-lg-3 mb-2">
-                                        <button type="button" class="btn btn-primary w-100 modal-button" data-value="firstname"
+                                        <button type="button" class="btn btn-primary w-100 modal-button btn-sm" data-value="firstname"
                                                 data-bs-toggle="modal"
                                                 data-bs-target="#account-edit-modal" id="account-edit-button-firstname">
                                             Firstname
                                         </button>
                                     </div>
                                     <div class="col-12 col-md-6 col-lg-3 mb-2">
-                                        <button type="submit" class="btn btn-primary w-100 modal-button" data-value="lastname"
+                                        <button type="submit" class="btn btn-primary w-100 modal-button btn-sm" data-value="lastname"
                                                 data-bs-toggle="modal"
                                                 data-bs-target="#account-edit-modal" id="account-edit-button-lastname">
                                             Lastname
                                         </button>
                                     </div>
                                     <div class="col-12 col-md-6 col-lg-3 mb-2">
-                                        <button type="submit" class="btn btn-primary w-100 modal-button" data-value="email"
+                                        <button type="submit" class="btn btn-primary w-100 modal-button btn-sm" data-value="email"
                                                 data-bs-toggle="modal"
                                                 data-bs-target="#account-edit-modal" id="account-edit-button-email">
                                             Email
                                         </button>
                                     </div>
                                     <div class="col-12 col-md-6 col-lg-3 mb-2">
-                                        <button type="submit" class="btn btn-primary w-100 modal-button" data-value="password"
+                                        <button type="submit" class="btn btn-primary w-100 modal-button btn-sm" data-value="password"
                                                 data-bs-toggle="modal"
                                                 data-bs-target="#account-edit-modal" id="account-edit-button-password">
                                             Password
@@ -51,7 +51,7 @@
                                 </div>
                                 <div class="row justify-content-center pt-3">
                                     <div class="col-12 col-md-6 col-lg-3 mb-2">
-                                        <button type="button" class="btn btn-secondary modal-button"
+                                        <button type="button" class="btn btn-secondary modal-button btn-sm"
                                                 data-bs-dismiss="modal"
                                                 aria-label="Close">Cancel
                                         </button>
@@ -105,10 +105,10 @@
                                     <div class="row align-items-center">
                                         <div class="col-lg-4 col-md-6 col-sm-12 mb-2">
                                             <div class="d-flex gap-3 justify-content-between pt-3">
-                                                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal"
+                                                <button type="button" class="btn btn-secondary btn-sm" data-bs-dismiss="modal"
                                                         aria-label="Close">Cancel
                                                 </button>
-                                                <button type="submit" class="btn btn-success" id="account-edit-modal-submit-button" disabled>Update</button>
+                                                <button type="submit" class="btn btn-success btn-sm" id="account-edit-modal-submit-button" disabled>Update</button>
                                             </div>
                                         </div>
                                     </div>
