+++ trip_calculator/trip_calculator/templates/trip_calculator/home_table/account_table.html
@@ -17,7 +17,7 @@
         </tr>
         </tbody>
         <caption class="pt-2">
-           <button type="button" class="btn btn-primary small-button" data-bs-toggle="modal" id="edit-account-button"
+           <button type="button" class="btn btn-primary btn-sm footer-tabel-button" data-bs-toggle="modal" id="edit-account-button"
                                     data-bs-target="#account-common-edit-modal">Edit account</button>
         </caption>
     </table>
