+++ trip_calculator/trip_calculator/templates/trip_calculator/home_table/cost_table.html
@@ -26,24 +26,46 @@
                     <ul class="dropdown-menu background-dropdown-color dropdown-list"
                         aria-labelledby="dropdownMenuButtonFriend"
                         style="width: 100px; max-height: 250px; white-space: nowrap; overflow-y: auto;">
-                        {% for name in cost.split_by %}
-                        <li class="ps-2 pe-2">{{ name.firstname }}</li>
+                        {% for data in cost.split_by %}
+                        <li class="ps-2 pe-2">
+
+                            {% if cost.who_pay.was_you %}
+
+                            <button type="button" class="btn p-0" data-bs-toggle="modal" id="edit-cost-button-{{ data.user_id }}"
+                                    data-bs-target="#cost-edit-modal" data-value-user_id="{{ data.user_id }}"
+                                    data-value-cost-id="{{ friend.user_id }}">
+                                <i class="bi bi-cash-coin me-1"
+                               style="color:{% if data.payment %}green {% else %}red {% endif %};"></i> {{ data.firstname }}
+                            </button>
+
+
+                            {% else %}
+                            <i class="bi bi-cash-coin me-1"
+                               style="color:{% if data.payment %}green {% else %}red {% endif %};"></i> {{ data.firstname }}
+                            {% endif %}
+
+                        </li>
                         {% endfor %}
                     </ul>
                 </div>
             </td>
             <td scope="col">{{ cost.cost }}</td>
             <td scope="col">{{ cost.unit_cost }}</td>
-            <td scope="col" style="color:{% if cost.return == '0.00' %}black{% elif cost.who_pay.was_you %}green{% else %}red{% endif %};">{{ cost.return }}</td>
+            <td scope="col"
+                style="color:{% if cost.return == '0.00' or cost.return == '0' %}black{% elif cost.who_pay.was_you %}green{% else %}red{% endif %};">
+                {{ cost.return }}
+            </td>
         </tr>
         {% endfor %}
 
         </tbody>
         <caption class="pt-2">
-            <a href="{% url 'add_cost_view' trip_id=trip.id %}" class="btn btn-success bi footer-tabel-button"> Add Cost</a>
+            <a href="{% url 'add_cost_view' trip_id=trip.id %}" class="btn btn-success bi btn-sm footer-tabel-button"> Add
+                Cost</a>
             <p class="h6">Person sum: {{ trip.totalCost }}</p>
             <p class="h6">Balance: {% if trip.balance.positive %}+ {% else %}- {% endif %}{{ trip.balance.amount }}</p>
         </caption>
     </table>
 </div>
 {% endfor %}
+{% include 'trip_calculator/home_table/cost_edit_modal.html' %}
