+++ trip_calculator/trip_calculator/templates/trip_calculator/home_table/trip_table.html
@@ -34,7 +34,7 @@
         {% endfor %}
         </tbody>
         <caption class="pt-2">
-            <a href="{% url 'create_trip_view' %}" class="btn btn-primary bi footer-tabel-button">Edit Trip</a>
+            <a href="{% url 'create_trip_view' %}" class="btn btn-primary bi btn-sm footer-tabel-button">Edit Trip</a>
         </caption>
     </table>
 </div>
